requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
DateTime==4.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.5
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
gspread==3.1.0
gspread-dataframe==3.0.3
gunicorn==19.9.0
httplib2==0.13.1
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.2
oauth2client==4.1.3
oauthlib==3.1.0
pandas==0.25.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
transliterate==1.10.2
uritemplate==3.0.0
urllib3==1.25.3
vk-api==11.5.0
zope.interface==4.6.0
psycopg2>=2.6.1
